Endpoint para buscar reteios de votos e scheduler para retaear votos se sessoes encerradas

+++ src/main/java/br/pereira/votous/api/v1/dto/VotoOutputDTO.java
@@ -10,7 +10,7 @@ public class VotoOutputDTO {
 
     private Long cdVoto;
     private Integer cdSessao;
-    private Integer cdPauta;
+    private Long cdPauta;
     private String cpf;
     private VotoEnum voto;
 }
