+++ src/main/java/br/pereira/votous/entity/VotoEntity.java
@@ -24,7 +24,7 @@ public class VotoEntity {
 
     @Column(name = "cd_pauta")
     @NotNull
-    private Integer cdPauta;
+    private Long cdPauta;
 
     @Column(name = "cd_sessao")
     @NotNull
